third_party/WebKit/LayoutTests/editing/pasteboard/innerText-inline-table.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="test1">hello<table style="display:inline-table"><tr><td>1</td></tr></table>world</div>
<div id="test2">foo <table style="display:inline-table"><tr><td>2</td></tr></table> bar</div>
<script>
test(() => {
  assert_equals('hello 1 world', document.getElementById('test1').innerText,
    'test1.innerText');
  assert_equals('foo  2  bar', document.getElementById('test2').innerText,
    'test2.innerText');
}, 'Checks that the text iterator is emitting a space before and after an inline table');
</script>
